requirements.txt
openiap-edge==0.0.28
